Multiple changes
Added initial {!a} conversion support
Moved compile_test_run.sh into test/CMakeLists.txt
Moved linting support to src/CMakeLists.txt
Printing uint8_t and int8_t as numbers!
Multiple changes related to linting.

+++ src/yio/yio/manip/print_count.c
@@ -16,7 +16,7 @@ int _yΩIO_print_count(yπio_printctx_t *t) {
 	if (writtencnt > INT_MAX) {
 		*arg = INT_MAX;
 	} else {
-		*arg = writtencnt;
+		*arg = (int)writtencnt;
 	}
 	return 0;
 }
